[stdlib] Deduplicate exception classes in non-JVM source sets
