+++ libraries/stdlib/wasm/src/kotlin/Exceptions.kt
@@ -5,96 +5,6 @@
 
 package kotlin
 
-public actual open class Error : Throwable {
-    public actual constructor() : super()
-    public actual constructor(message: String?) : super(message)
-    public actual constructor(message: String?, cause: Throwable?) : super(message, cause)
-    public actual constructor(cause: Throwable?) : super(cause)
-}
-
-public actual open class Exception : Throwable {
-    public actual constructor() : super()
-    public actual constructor(message: String?) : super(message)
-    public actual constructor(message: String?, cause: Throwable?) : super(message, cause)
-    public actual constructor(cause: Throwable?) : super(cause)
-}
-
-public actual open class RuntimeException : Exception {
-    public actual constructor() : super()
-    public actual constructor(message: String?) : super(message)
-    public actual constructor(message: String?, cause: Throwable?) : super(message, cause)
-    public actual constructor(cause: Throwable?) : super(cause)
-}
-
-public actual open class IllegalArgumentException : RuntimeException {
-    public actual constructor() : super()
-    public actual constructor(message: String?) : super(message)
-    public actual constructor(message: String?, cause: Throwable?) : super(message, cause)
-    public actual constructor(cause: Throwable?) : super(cause)
-}
-
-public actual open class IllegalStateException : RuntimeException {
-    public actual constructor() : super()
-    public actual constructor(message: String?) : super(message)
-    public actual constructor(message: String?, cause: Throwable?) : super(message, cause)
-    public actual constructor(cause: Throwable?) : super(cause)
-}
-
-public actual open class IndexOutOfBoundsException : RuntimeException {
-    public actual constructor() : super()
-    public actual constructor(message: String?) : super(message)
-}
-
-public actual open class ConcurrentModificationException : RuntimeException {
-    public actual constructor() : super()
-    public actual constructor(message: String?) : super(message)
-    public actual constructor(message: String?, cause: Throwable?) : super(message, cause)
-    public actual constructor(cause: Throwable?) : super(cause)
-}
-
-public actual open class UnsupportedOperationException : RuntimeException {
-    public actual constructor() : super()
-    public actual constructor(message: String?) : super(message)
-    public actual constructor(message: String?, cause: Throwable?) : super(message, cause)
-    public actual constructor(cause: Throwable?) : super(cause)
-}
-
-
-public actual open class NumberFormatException : IllegalArgumentException {
-    public actual constructor() : super()
-    public actual constructor(message: String?) : super(message)
-}
-
-
-public actual open class NullPointerException : RuntimeException {
-    public actual constructor() : super()
-    public actual constructor(message: String?) : super(message)
-}
-
-public actual open class ClassCastException : RuntimeException {
-    public actual constructor() : super()
-    public actual constructor(message: String?) : super(message)
-}
-
-public actual open class AssertionError : Error {
-    public actual constructor() : super()
-    public constructor(message: String?) : super(message)
-    public actual constructor(message: Any?) : super(message?.toString(), message as? Throwable)
-    @SinceKotlin("1.4")
-    public actual constructor(message: String?, cause: Throwable?) : super(message, cause)
-}
-
-public actual open class NoSuchElementException : RuntimeException {
-    public actual constructor() : super()
-    public actual constructor(message: String?) : super(message)
-}
-
-@SinceKotlin("1.3")
-public actual open class ArithmeticException : RuntimeException {
-    public actual constructor() : super()
-    public actual constructor(message: String?) : super(message)
-}
-
 @Suppress("EXPECT_ACTUAL_INCOMPATIBLE_VISIBILITY")
 internal actual open class NoWhenBranchMatchedException : RuntimeException {
     public actual constructor() : super()
